menu_detailをファイル分割
